requirements.txt
absl-py==0.2.2
amqp==2.3.2
asn1crypto==0.24.0
astor==0.6.2
billiard==3.5.0.3
bleach==1.5.0
celery==4.2.0
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
click==6.7
common==0.1.2
cryptography==2.2.2
cycler==0.10.0
dnspython==1.15.0
email-validator==1.0.2
Flask==1.0.2
Flask-MySQL==1.4.0
Flask-SQLAlchemy==2.3.2
Flask-Validator==1.2.3
gast==0.2.0
grpcio==1.12.1
h5py==2.8.0
html5lib==0.9999999
idna==2.7
isbnlib==3.6.8
iso3166==0.8
itsdangerous==0.24
Jinja2==2.10
Keras==2.2.0
Keras-Applications==1.0.2
Keras-Preprocessing==1.0.1
kiwisolver==1.0.1
kombu==4.2.1
Markdown==2.6.11
MarkupSafe==1.0
matplotlib==2.2.2
meld3==1.0.2
numpy==1.14.5
Pillow
protobuf==3.6.0
py-moneyed==0.7.0
pyactiveresource==2.1.2
pycparser==2.18
PyMySQL==0.9.0
pyparsing==2.2.0
python-dateutil==2.7.3
pytz==2017.2
PyYAML==3.12
redis==2.10.6
requests==2.19.1
scikit-learn==0.19.1
scipy==1.1.0
ShopifyAPI==3.1.0
six==1.11.0
sklearn==0.0
SQLAlchemy==1.2.8
git+https://github.com/Supervisor/supervisor
tensorboard==1.8.0
tensorflow
termcolor==1.1.0
tqdm==4.23.4
urllib3==1.23
vine==1.1.4
virtualenv==16.0.0
Werkzeug==0.14.1
